+++ public/style.css
@@ -29,6 +29,11 @@ body {
   padding: 10px;
   font-family: Verdana, sans-serif;
 }
+
+
+
+
+/* ADD DATA PANEL */
 body #add-data-panel {
   display: flex; flex-direction: column; justify-content: center; align-items: center; gap: 20px;
   user-select: none;
@@ -50,46 +55,22 @@ body #add-data-panel span {
 body #add-data-panel:hover img {
   opacity: 1;
 }
+
+
+
+
+/* MAIN */
 body main {
   /* width: 100vw; */ height: 100%;
   display: flex; justify-content: flex-start; align-items: flex-start; gap: 10px;
   overflow: auto;
   padding: 100px;
 }
-/* body::-webkit-scrollbar {
-  display: none;
-} */
-body main .link-item {
-  width: 100px; height: 100px;
-  display: inline-block;
-  /* margin: 5px; */
-  position: relative;
-}
-body main .link-item .link-anchor {
-  background: #ffffff; background-size: 100%; background-position: center; background-repeat: no-repeat;
-  width: 100%; height: 100%;
-  border-radius: 5px;
-  display: flex; justify-content: flex-start; align-items: flex-start;
-  text-decoration: none; color: #333333;
-  /* padding: 5px; */
-  box-shadow: 0 1px 2px rgba(0,0,0,0.12), 0 1px 4px rgba(0,0,0,0.24);
-  transition: all 0.1s ease-in;
-}
-body main .link-item .link-anchor:hover {
-  box-shadow: 0 4px 8px rgba(0,0,0,0.12), 0 4px 16px rgba(0,0,0,0.24);
-  animation: push-in 0.3s ease-in-out 1;
-}
-/* body main .services .link-item .link-anchor img {
-  width: 100%; height: 100%;
-  border-radius: 5px;
-} */
-body main  .disabled-link .link-anchor {
-  pointer-events: none;
-  opacity: .1;
-}
-body main .fixed-tool-button {
-  position: absolute;
-}
+
+
+
+
+/* FORM */
 body main .tool-button {
   display: block;
   width: 20px; height: 20px;
@@ -162,6 +143,10 @@ body #form {
   transition: border-color 0.3s ease-in-out;
 }
 
+
+
+
+/* ANIMATIONS */
 @keyframes pop-in {
   70% {
     transform: scale(1.2);
